fix: using a better component structure
