--- src/components/AccountCard/AccountCardContainer.types.tsx
@@ -1,3 +1,0 @@
-export type Props = {
-  children: any
-}
