+++ src/components/HomePage/AccountCard/AccountCard.types.tsx
@@ -1,0 +1,25 @@
+export enum NetworkType {
+  MATIC = 'matic',
+  ETHEREUM = 'ethereum'
+}
+
+export enum TransactionType {
+  DEPOSIT = 'deposit',
+  WITHDRAWAL = 'withdrawal'
+}
+
+export enum TransactionStatus {
+  CONFIRMED = 'confirmed',
+  PENDING = 'pending'
+}
+
+export type Transaction = {
+  type: TransactionType
+  status: TransactionStatus
+  amount: number
+  description: string
+}
+
+export type Props = {
+  children: any
+}
