+++ src/components/HomePage/AccountCard/AccountCardHeader/AccountCardHeader.types.tsx
@@ -1,0 +1,7 @@
+import { NetworkType } from '../AccountCard.types'
+
+export type Props = {
+  type: NetworkType
+  title: string
+  amount: number
+}
