+++ src/components/HomePage/AccountCard/AccountCardHeader/index.ts
@@ -1,0 +1,2 @@
+import AccountCardHeader from './AccountCardHeader'
+export { AccountCardHeader }
