+++ src/components/HomePage/AccountCard/AccountTransactions/AccountTransaction/AccountTransaction.types.tsx
@@ -1,0 +1,3 @@
+import { Transaction } from '../../AccountCard.types'
+
+export type Props = Transaction
