+++ src/components/HomePage/AccountCard/AccountTransactions/AccountTransaction/index.ts
@@ -1,0 +1,2 @@
+import AccountTransaction from './AccountTransaction'
+export { AccountTransaction }
