+++ src/components/HomePage/AccountCard/AccountTransactions/AccountTransactions.types.tsx
@@ -1,0 +1,5 @@
+import { Transaction } from '../AccountCard.types'
+
+export type Props = {
+  transactions: Transaction[]
+}
