+++ src/components/HomePage/AccountCard/AccountTransactions/index.ts
@@ -1,0 +1,2 @@
+import AccountTransactions from './AccountTransactions'
+export { AccountTransactions }
